Split ``Index{Entry}``
